+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Updated
+- Added more cached parameters into Scroll providers. [Trello 1854](https://trello.com/c/Sdzfuue2)
+
 ## [3.198.0]
 ### Updated
 - Closing test instead of aborting them when getAllTestsResults called without closing eyes by the user. [Trello 2493](https://trello.com/c/YS3vLGxd)